fix the runtime issue

+++ kernel/script/img-write-osx.sh
@@ -2,8 +2,18 @@ if [ -f "disk1.vhd" ]; then
     mv disk1.vhd disk1.dmg
 fi
 
+if [ ! -f "disk1.dmg" ]; then
+    echo "error: no disk1.vhd, download it first!!!"
+    exit -1
+fi
+
 if [ -f "disk2.vhd" ]; then
     mv disk2.vhd disk2.dmg
+fi
+
+if [ ! -f "disk2.dmg" ]; then
+    echo "error: no disk2.vhd, download it first!!!"
+    exit -1
 fi
 
 export DISK1_NAME=disk1.dmg
@@ -19,16 +29,16 @@ dd if=kernel.elf of=$DISK1_NAME bs=512 conv=notrunc seek=100
 
 # 写应用程序init，临时使用
 # dd if=init.elf of=$DISK1_NAME bs=512 conv=notrunc seek=5000
-# dd if=shell.elf of=$DISK1_NAME bs=512 conv=notrunc seek=5000
+dd if=shell.elf of=$DISK1_NAME bs=512 conv=notrunc seek=5000
 
 # 写应用程序，使用系统的挂载命令
-export DISK2_NAME=disk2.dmg
-export TARGET_PATH=mp
-rm $TARGET_PATH
-hdiutil attach $DISK2_NAME -mountpoint $TARGET_PATH
+# export DISK2_NAME=disk2.dmg
+# export TARGET_PATH=mp
+# rm $TARGET_PATH
+# hdiutil attach $DISK2_NAME -mountpoint $TARGET_PATH
 # cp -v init.elf $TARGET_PATH/init
 # cp -v shell.elf $TARGET_PATH
 # cp -v loop.elf $TARGET_PATH/loop
 # cp -v snake.elf $TARGET_PATH/snake
-cp -v *.elf $TARGET_PATH
-hdiutil unmount $TARGET_PATH -verbose
+# cp -v *.elf $TARGET_PATH
+# hdiutil unmount $TARGET_PATH -verbose
